Refactor table UI boundaries
