+++ apps/web/src/ui/tables/balances-table/index.ts
@@ -1,0 +1,1 @@
+export { BalancesTable } from "./BalancesTable";
