+++ apps/web/src/ui/tables/budget-table/fx-breakdown/logic.ts
