+++ apps/web/src/ui/tables/budget-table/index.ts
@@ -1,12 +1,2 @@
-export {
-  BudgetTableHeader,
-  type BudgetTableHeaderProps,
-} from "./BudgetTableHeader";
-export {
-  BudgetDirectionSection,
-  type BudgetDirectionSectionProps,
-} from "./BudgetDirectionSection";
-export {
-  BudgetDerivedSection,
-  type BudgetDerivedSectionProps,
-} from "./BudgetDerivedSection";
+export { BudgetTable } from "./BudgetTable";
+export type { BudgetTableProps } from "./useBudgetTableController";
