+++ apps/web/src/ui/tables/data-table/index.ts
@@ -1,0 +1,11 @@
+export { DataTable } from "./DataTable";
+export { useInfiniteScroll } from "./useInfiniteScroll";
+export { useTableSort } from "./useTableSort";
+export type {
+  ColumnDef,
+  DataTableProps,
+  InfiniteScrollState,
+  PageResult,
+  SortEntry,
+  SortState,
+} from "./types";
