+++ apps/web/src/ui/tables/shared/TableUi.module.css
@@ -68,6 +68,10 @@
   font-size: 12px;
 }
 
+.error {
+  color: #c0392b;
+}
+
 .stale {
   color: #c0392b;
 }
@@ -251,6 +255,17 @@
 .scrollSentinel {
   min-height: 1px;
   padding: 8px 0;
+}
+
+.loadingEdge {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-width: 150px;
+  flex-shrink: 0;
+  font-size: 12px;
+  color: var(--muted);
+  white-space: nowrap;
 }
 
 .overlayBackdrop {
