+++ apps/web/src/ui/tables/transactions-table/filters/FilterMultiSelect.tsx
@@ -7,7 +7,7 @@ import { useTranslation } from "react-i18next";
 
 import { cn } from "@/lib/cn";
 
-import styles from "./TableUi.module.css";
+import styles from "../../shared/TableUi.module.css";
 
 type Rect = Readonly<{ top: number; left: number; width: number; height: number }>;
