+++ apps/web/src/ui/tables/transactions-table/index.ts
@@ -1,0 +1,3 @@
+export { DrillDownPanel } from "./DrillDownPanel";
+export { TransactionsRawTable } from "./TransactionsRawTable";
+export type { DrillDownFilter } from "./DrillDownPanel";
